feat: return nodeOperatorFeeRate as uint256

+++ si-contracts/0.8.25/vaults/VaultViewer.sol
@@ -16,7 +16,7 @@ contract VaultViewer {
         VaultHub.VaultRecord record;
         uint256 totalValue;
         uint256 liabilityStETH;
-        int128 nodeOperatorFeeRate;
+        uint256 nodeOperatorFeeRate;
         uint256 accruedFee;
         bool isReportFresh;
         LazyOracle.QuarantineInfo quarantineInfo;
@@ -180,7 +180,7 @@ contract VaultViewer {
         ILido lido = VAULT_HUB.LIDO();
         VaultHub.VaultConnection memory connection = VAULT_HUB.vaultConnection(vault);
         VaultHub.VaultRecord memory record = VAULT_HUB.vaultRecord(vault);
-        int128 nodeOperatorFeeRate = _getNodeOperatorFeeRate(connection.owner);
+        uint256 nodeOperatorFeeRate = _getNodeOperatorFeeRate(connection.owner);
         uint256 accruedFee = _getAccruedFee(connection.owner);
         LazyOracle.QuarantineInfo memory quarantineInfo = LAZY_ORACLE.vaultQuarantine(vault);
 
@@ -327,13 +327,13 @@ contract VaultViewer {
     /// @dev Uses low-level staticcall to avoid reverting when the method is missing or the address is an EOA
     /// @param owner The address of the vault owner (can be either a contract or an EOA)
     /// @return fee The decoded fee value if present, otherwise 0
-    function _getNodeOperatorFeeRate(address owner) internal view returns (int128 fee) {
+    function _getNodeOperatorFeeRate(address owner) internal view returns (uint256 fee) {
         if (_isContract(owner)) {
             // if dashboard contract and have feeRate method
             (bool success, bytes memory result) = owner.staticcall(abi.encodeWithSignature("feeRate()"));
             // Check ensures safe decoding — avoids abi.decode revert on short return data
             if (success && result.length >= 32) {
-                fee = abi.decode(result, (int128));
+                fee = abi.decode(result, (uint256));
             }
         }
     }
